# Improve performance of access control query enhancer