| Q | A
| --- | ---
| Bug fix? | no 
| New feature? | no 
| BC breaks? | no 
| Deprecations? | no  <!-- if yes add them to the UPGRADE.md file -->
| Fixed tickets | fixes # <!-- add issue number here e.g.: #5730 -->
| Related issues/PRs | # <!-- add issue or PR number here e.g.: #5730 -->
| License | MIT
| Documentation PR | sulu/sulu-docs# <!-- add docs PR number here e.g.: sulu/sulu-docs#615 -->

#### What's in this PR?

Improve performance of access control query enhancer.

The change removes the subquery and selects all IDs which are not accessible by the user and exclude them from the real result.

#### Why?

Currently the queries with a lot of medias are taken a lot of time:

This will improve the list query `7462.32 ms` -> `2.34 ms + 2.64 ms`. 
The count query will also improve from `4872.34 ms`-> `2.83 ms + 635.06 ms`.

So its an over all improvement from `12361.82 ms` -> `659.44 ms`.

To implement this performance improvement we use an own query to load the restricted (not accessible) ids. This is faster as embedding the query as subquery (as it was before). Also loading the "accessible" ids would be more performance intense because sulu has more "accessible" entities that not accessible. Optimized embedded queries are mostly not compatible for MySQL 5.7 because of restrictions.

 As long as we dont have thousands of not "accessible" ids this approach should be faster.