@alexander-schranz 2021-11-17
@wachterjohannes can you add the stats and differents here as comment on the 2 tested databases (before / after).